+++ aanalytics2/aanalytics14.py
@@ -4,6 +4,7 @@ from aanalytics2 import config, connector
 from typing import Union
 from copy import deepcopy
 import pandas as pd
+import json
 
 class Report:
 
@@ -29,6 +30,9 @@ class Report:
     
     def __repr__(self):
         return self.DataFrame.to_string()
+    
+    def getDataFrame(self):
+        return self.DataFrame
 
 class LegacyAnalytics:
     """
@@ -168,4 +172,155 @@ class LegacyAnalytics:
                                     data.get('report').get('period'))
                 return class_data
             return df
-            
+
+class ReportBuilder14:
+
+    """
+    Build a report for the Legacy Analytics 1.4 API Report capability. 
+    """
+
+    def __init__(self,template:Union[dict,'ReportBuilder14']=None)->None:
+        """
+        Construct a report based information.
+        Arguments:
+            template : OPTIONAL : If you want to use a template of a report to not start from scratch.
+        """
+        report = {'reportDescription':{'reportSuiteID':None,'elements':[],'metrics':[]}}
+        if template is not None:
+            if type(template) == ReportBuilder14:
+                report = template.to_dict()
+            elif type(template) == dict:
+                if 'reportDescription' in template.keys():
+                    template = template['reportDescription']
+                for key in template.keys():
+                    report['reportDescription'][key] = template[key]
+                if 'metrics' not in report['reportDescription'].keys():
+                    report['metrics'] = []
+                if 'elements' not in report['reportDescription'].keys():
+                    report['elements'] = []
+        self.report = report
+    
+    def __str__(self):
+        return json.dumps(self.report,indent=2)
+    
+    def __repr__(self):
+        return json.dumps(self.report,indent=2)
+    
+    def to_dict(self):
+        """
+        Return the report as dictionary
+        """
+        return deepcopy(self.report)
+
+    def setReportSuite(self,rsid:str):
+        """
+        Add a reportSuite
+        Argument:
+            rside : REQUIRED : The reportSuite ID
+        """
+        self.report["reportDescription"]['reportSuiteID'] = rsid
+    
+    def setSource(self,source:str="warehouse"):
+        """
+        In case yu want to define data warehouse report per example
+        Argument: 
+            source : REQUIRED : The source, by default "warehouse"
+        """
+        self.report["reportDescription"]["source"] = source
+    
+    def removeSource(self):
+        """
+        remove the source key in the report.
+        """
+        if 'source' in self.report['reportDescription'].keys():
+            del self.report['reportDescription']['source']
+    
+    def setDate(self,date:str=None,dateFrom:str=None,dateTo:str=None):
+        """
+        Set the date for the report. You can use either date or the dateFrom and dateTo parameter.
+        Arguments:
+            date : OPTIONAL : If requesting a report for a single period, the day/month/year that you want to run the report for. If you use this field, do not use dateFrom or dateTo. If all date fields are omitted, defaults to the current day. This field supports the following formats:
+                    Use YYYY for the desired year.
+                    Use YYYY-MM for the desired month.
+                    Use YYYY-MM-DD for the desired day.
+            dateFrom : OPTIONAL : The starting date range. If you use this field, also include dateTo and do not use date. The date format is YYYY-MM-DD. The month and day designators can be omitted if you want a monthly or yearly report
+            dateTo : OPTIONAL : The ending date range (inclusive). If you use this field, also include dateFrom and do not use date. The date format is YYYY-MM-DD. The month and day designators can be omitted if you want a monthly or yearly report
+        """
+        
+        if date is not None:
+            if 'dateTo' in self.report['reportDescription'].keys():
+                del self.report['reportDescription']["dateTo"]
+            if 'dateFrom' in self.report['reportDescription'].keys():
+                del self.report['reportDescription']["dateFrom"]
+            self.report['reportDescription']['date'] = date
+        else:
+            if dateTo is None or dateFrom is None:
+                raise Exception("Require a date parameter to be used")
+            else:
+                if "date" in self.report['reportDescription'].keys():
+                    del self.report['reportDescription']["date"]
+                self.report['reportDescription']['dateFrom'] = dateFrom
+                self.report['reportDescription']['dateTo'] = dateTo
+
+    def setGranularity(self,dateGranularity:str=None):
+        """
+        Specifies the date granularity used to display report data (trended reports). If this field is omitted, data is aggregated across the entire date range (ranked reports).
+        Argument:
+            dateGranularity : REQUIRED : Supported values are:
+                minute: Real-Time reports only. Displays report data by minute. Include an integer between 1 - 60 after a semicolon to increase the minute interval. For example, use minute:3 to aggregate data in 3-minute intervals.
+                hour: Displays report data by hour.
+                day: Displays report data by day.
+                week: Displays report data by week.
+                month: Displays report data by month.
+                quarter: Displays report data by quarter.
+                year: Displays report data by year.
+        """
+        self.report['reportDescription']['dateGranularity'] = dateGranularity
+
+    def removeGranularity(self):
+        """
+        Remove the granularity component
+        """
+        if 'dateGranularity' in self.report['reportDescription'].keys():
+            del self.report['reportDescription']['dateGranularity']
+    
+    def addMetric(self,metricId:str=None,**kwargs):
+        """
+        Add a specific metric
+        Argument: 
+            metricId : REQUIRED : The metric ID to be used
+            kwargs : OPTIONAL : any additional attribute you want to add
+        """
+        dict_metrics = {
+            "id":metricId
+        }
+        if len(kwargs)>0:
+            dict_metrics.update(**kwargs)
+        self.report['reportDescription']['metrics'].append(dict_metrics)
+
+    def addElement(self,elementId:str=None,**kwargs):
+        """
+        Add a specific metric
+        Argument: 
+            elementId : REQUIRED : The element ID to be used
+            kwargs : OPTIONAL : any additional attribute you want to add
+        """
+        dict_element = {
+            "id":elementId
+        }
+        if len(kwargs)>0:
+            dict_element.update(**kwargs)
+        self.report['reportDescription']['elements'].append(dict_element)
+
+
+    def setEncoding(self,encoding:str="utf8"):
+        """
+        Supported values include utf8 or base64.
+        Arguments:
+            encodning : OPTIONAL : either utf-8 or bas64.
+                utf8: Filters out invalid UTF-8 characters in the request and response.
+                base64: Treats the entire request, including element names, search/pathing filters, special keywords, and dates, as if they are base64 encoded.
+        """
+        self.report['reportDescription']['elementDataEncoding'] = encoding
+
+